@@ -2,7 +2,7 @@
   "name": "@minimal/material-kit-react",
   "author": "minimals.cc",
   "licence": "MIT",
-  "version": "1.4.0",
+  "version": "1.5.0",
   "private": false,
   "scripts": {
     "start": "react-scripts start",
@@ -10,7 +10,10 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "lint": "eslint --ext .js,.jsx ./src",
-    "lint:fix": "eslint --fix --ext .js,.jsx ./src"
+    "lint:fix": "eslint --fix --ext .js,.jsx ./src",
+    "clear-all": "rm -rf build node_modules",
+    "re-start": "rm -rf build node_modules && yarn install && yarn start",
+    "re-build": "rm -rf build node_modules && yarn install && yarn build"
   },
   "eslintConfig": {
     "extends": [
@@ -36,45 +39,50 @@
     ]
   },
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@faker-js/faker": "^6.1.2",
-    "@iconify/react": "^3.2.1",
-    "@mui/lab": "^5.0.0-alpha.77",
-    "@mui/material": "^5.6.1",
-    "@testing-library/jest-dom": "^5.16.2",
-    "apexcharts": "^3.35.0",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@faker-js/faker": "^7.3.0",
+    "@hookform/resolvers": "^2.9.3",
+    "@iconify/react": "^3.2.2",
+    "@mui/lab": "^5.0.0-alpha.88",
+    "@mui/material": "^5.8.6",
+    "@testing-library/jest-dom": "^5.16.4",
+    "apexcharts": "^3.35.3",
     "change-case": "^4.1.2",
     "date-fns": "^2.28.0",
-    "formik": "^2.2.9",
-    "history": "^5.2.0",
+    "history": "^5.3.0",
     "lodash": "^4.17.21",
     "numeral": "^2.0.6",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-apexcharts": "^1.4.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
+    "react-hook-form": "^7.33.1",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "^5.0.0",
-    "simplebar": "^5.3.6",
-    "simplebar-react": "^2.3.6",
+    "react-scripts": "^5.0.1",
+    "simplebar": "^5.3.8",
+    "simplebar-react": "^2.4.1",
     "web-vitals": "^2.1.4",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/eslint-parser": "^7.17.0",
-    "eslint": "^8.9.0",
+    "@babel/core": "^7.18.6",
+    "@babel/eslint-parser": "^7.18.2",
+    "@svgr/webpack": "^6.2.1",
+    "eslint": "^8.19.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.4.0",
-    "eslint-config-react-app": "^7.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "^2.5.1"
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.7.1"
+  },
+  "overrides": {
+    "@svgr/webpack": "^6.2.1"
   }
 }
