@@ -1,7 +1,7 @@
 // scroll bar
 import 'simplebar/src/simplebar.css';
 
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
 
@@ -12,13 +12,14 @@ import reportWebVitals from './reportWebVitals';
 
 // ----------------------------------------------------------------------
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
   <HelmetProvider>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </HelmetProvider>,
-  document.getElementById('root')
+  </HelmetProvider>
 );
 
 // If you want to enable client cache, register instead.
