@@ -58,7 +58,7 @@ function NewsItem({ news }) {
       <Box component="img" alt={title} src={image} sx={{ width: 48, height: 48, borderRadius: 1.5, flexShrink: 0 }} />
 
       <Box sx={{ minWidth: 240, flexGrow: 1 }}>
-        <Link color="inherit" variant="subtitle2" noWrap>
+        <Link color="inherit" variant="subtitle2" underline="hover" noWrap>
           {title}
         </Link>
 
