@@ -5,9 +5,9 @@ export default function Autocomplete(theme) {
     MuiAutocomplete: {
       styleOverrides: {
         paper: {
-          boxShadow: theme.customShadows.z20
-        }
-      }
-    }
+          boxShadow: theme.customShadows.z20,
+        },
+      },
+    },
   };
 }
