@@ -7,20 +7,16 @@ export default function IconButton(theme) {
         {
           props: { color: 'default' },
           style: {
-            '&:hover': { backgroundColor: theme.palette.action.hover }
-          }
+            '&:hover': { backgroundColor: theme.palette.action.hover },
+          },
         },
         {
           props: { color: 'inherit' },
           style: {
-            '&:hover': { backgroundColor: theme.palette.action.hover }
-          }
-        }
+            '&:hover': { backgroundColor: theme.palette.action.hover },
+          },
+        },
       ],
-
-      styleOverrides: {
-        root: {}
-      }
-    }
+    },
   };
 }
