@@ -7,29 +7,29 @@ export default function Lists(theme) {
         root: {
           color: 'inherit',
           minWidth: 'auto',
-          marginRight: theme.spacing(2)
-        }
-      }
+          marginRight: theme.spacing(2),
+        },
+      },
     },
     MuiListItemAvatar: {
       styleOverrides: {
         root: {
           minWidth: 'auto',
-          marginRight: theme.spacing(2)
-        }
-      }
+          marginRight: theme.spacing(2),
+        },
+      },
     },
     MuiListItemText: {
       styleOverrides: {
         root: {
           marginTop: 0,
-          marginBottom: 0
+          marginBottom: 0,
         },
         multiline: {
           marginTop: 0,
-          marginBottom: 0
-        }
-      }
-    }
+          marginBottom: 0,
+        },
+      },
+    },
   };
 }
