@@ -4,14 +4,14 @@ export default function Paper() {
   return {
     MuiPaper: {
       defaultProps: {
-        elevation: 0
+        elevation: 0,
       },
 
       styleOverrides: {
         root: {
-          backgroundImage: 'none'
-        }
-      }
-    }
+          backgroundImage: 'none',
+        },
+      },
+    },
   };
 }
