@@ -5,12 +5,12 @@ export default function Tooltip(theme) {
     MuiTooltip: {
       styleOverrides: {
         tooltip: {
-          backgroundColor: theme.palette.grey[800]
+          backgroundColor: theme.palette.grey[800],
         },
         arrow: {
-          color: theme.palette.grey[800]
-        }
-      }
-    }
+          color: theme.palette.grey[800],
+        },
+      },
+    },
   };
 }
