@@ -5,12 +5,12 @@ export default function Typography(theme) {
     MuiTypography: {
       styleOverrides: {
         paragraph: {
-          marginBottom: theme.spacing(2)
+          marginBottom: theme.spacing(2),
         },
         gutterBottom: {
-          marginBottom: theme.spacing(1)
-        }
-      }
-    }
+          marginBottom: theme.spacing(1),
+        },
+      },
+    },
   };
 }
